+++ ChangeLog.md
@@ -10,6 +10,9 @@
   destinations. ([#533](https://github.com/davep/blogmore/pull/533))
 - Added a related posts system.
   ([#538](https://github.com/davep/blogmore/pull/538))
+- Added a `drafts` command line command to list the paths of all posts
+  marked as being a draft.
+  ([#539](https://github.com/davep/blogmore/pull/539))
 
 ## v2.29.0